refactor: structure files by moving to correct folder
