+++ src/sections/skills/components/carousel/utils/getMockCarousel.ts
